Rename message-queue-v2 to queue-manager
It's one the primary service class is called. It's what we decorate the server and request objects with.

Hopefully calling it this will make a clearer link between the reference and the source when we come to be looking at code in the future.
