+++ src/lib/queue-manager/QueueManager.js
