+++ src/modules/billing/jobs/process-charge-version-year.js
@@ -6,7 +6,7 @@ const JOB_NAME = 'billing.process-charge-version-year'
 const { logger } = require('../../../logger')
 const { jobName: prepareTransactionsJobName } = require('./prepare-transactions')
 
-const messageQueue = require('../../../lib/message-queue-v2')
+const queueManager = require('../../../lib/queue-manager')
 
 const { BATCH_ERROR_CODE } = require('../../../lib/models/batch')
 const batchJob = require('./lib/batch-job')
@@ -36,7 +36,7 @@ const handler = async job => {
     return new Promise((resolve, reject) => {
       child.on('message', msg => {
         if (msg.complete === true && msg.batchId) {
-          messageQueue.getQueueManager().add(prepareTransactionsJobName, msg.batchId)
+          queueManager.getQueueManager().add(prepareTransactionsJobName, msg.batchId)
         } else if (msg.error) {
           logger.error(msg.error)
           reject(msg.error)
