+++ test/lib/services/invoice-account-addresses-service.test.js
@@ -21,20 +21,20 @@ const InvoiceAccountAddress = require('../../../src/lib/models/invoice-account-a
 const invoiceAccountsConnector = require('../../../src/lib/connectors/crm-v2/invoice-accounts')
 const invoiceAccountAddressesConnector = require('../../../src/lib/connectors/crm-v2/invoice-account-addresses')
 
-const messageQueue = require('../../../src/lib/message-queue-v2')
+const queueManager = require('../../../src/lib/queue-manager')
 const { jobNames } = require('../../../src/lib/constants')
 
 experiment('modules/billing/services/invoice-account-addreses-service', () => {
-  let messageQueueStub
+  let queueManagerStub
 
   beforeEach(async () => {
-    messageQueueStub = {
+    queueManagerStub = {
       add: sandbox.stub()
     }
     sandbox.stub(invoiceAccountsConnector, 'createInvoiceAccountAddress')
     sandbox.stub(invoiceAccountAddressesConnector, 'deleteInvoiceAccountAddress').resolves()
     sandbox.stub(invoiceAccountAddressesConnector, 'patchInvoiceAccountAddress').resolves()
-    sandbox.stub(messageQueue, 'getQueueManager').returns(messageQueueStub)
+    sandbox.stub(queueManager, 'getQueueManager').returns(queueManagerStub)
   })
 
   afterEach(async () => {
@@ -86,7 +86,7 @@ experiment('modules/billing/services/invoice-account-addreses-service', () => {
     })
 
     test('publishes a job to update the address in the CM', async () => {
-      expect(messageQueueStub.add.calledWith(
+      expect(queueManagerStub.add.calledWith(
         jobNames.updateCustomerAccount, invoiceAccountId
       )).to.be.true()
     })
